reworked the effect api a little and made it possible for skills to override effects in their config

+++ src/main/java/de/raidcraft/skills/api/combat/effect/PeriodicEffect.java
@@ -7,7 +7,7 @@ public interface PeriodicEffect<S, T> extends TimedEffect<S, T>, Runnable {
 
     public int getTaskId();
 
-    public int getDelay();
+    public long getDelay();
 
-    public int getInterval();
+    public long getInterval();
 }
